+++ RELEASES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## [Unreleased]
+
+### Added
+
+* powercap-info: control type name no longer required - will print info for all control types if not set
+
+
 ## [v0.5.0] - 2021-12-09
 
 ### Added
@@ -138,6 +145,7 @@
 
 * Initial public release
 
+[Unreleased]: https://github.com/powercap/powercap/compare/v0.5.0...HEAD
 [v0.5.0]: https://github.com/powercap/powercap/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/powercap/powercap/compare/v0.3.1...v0.4.0
 [v0.3.1]: https://github.com/powercap/powercap/compare/v0.3.0...v0.3.1